feat: decoupling of species Edit page
